✅ Pow(x, n) | Medium | C++ | Runtime: 0 ms | Memory: 8 MB | 2026-09-02

+++ leetcode/medium/0050-powx-n/solution.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long myPow(double x, int n) {
+    double myPow(double x, int n) {
         if(n==0){
             return 1;
         }
